+++ GreyGui.Core/src/Data/FontInfo.cs
@@ -7,7 +7,7 @@ public class FontInfo
     public Typeface Typeface { get; private set; }
     public Dictionary<char, GlyphInfo> GlyphInfoMap { get; private set; } = [];
     public float FontSizeOneSpaceWidth { get; set; }
-    public Dictionary<char, int> GlyphInfoIndexMap { get; private set; } = [];
+    public Dictionary<char, ushort> GlyphInfoIndexMap { get; private set; } = [];
 
     public FontInfo(string ttfPath)
     {
@@ -18,5 +18,5 @@ public class FontInfo
             FontSizeOneSpaceWidth = Typeface.GetHAdvanceWidthFromGlyphIndex(spaceGlyphIndex);
         }
     }
-    public int GetCharIndex(char c) { return GlyphInfoIndexMap[c]; }
+    public ushort GetCharIndex(char c) { return GlyphInfoIndexMap[c]; }
 }
